Alteration for version passing
